rpi5: Minimise direct stream to cache on store settings
Setting all these to zero gave the best geekbench scores
+0.7% single-core, and +3.35% multi-core.

WS_THR_L2, [25:24] Threshold for direct stream to L2 cache on store. The possible values are:
00 256B.
01 4KB. This is the reset value.
10 8KB.
11 Disables direct stream to L2 cache on store.

WS_THR_L3, [23:22] Threshold for direct stream to L3 cache on store. The possible values are:
00 768B.
01 16KB. This is the reset value.
10 32KB.
11 Disables direct stream to L3 cache on store.

WS_THR_L4, [21:20] Threshold for direct stream to L4 cache on store. The possible values are:
00 16KB.
01 64KB. This is the reset value.
10 128KB.
11 Disables direct stream to L4 cache on store.

WS_THR_DRAM, [19:18] Threshold for direct stream to DRAM on store. The possible values are:
00 64KB.
01 1MB, for memory designated as outer-allocate. This is the reset value.
10 1MB, allocating irrespective of outer-allocation designation.
11 Disables direct stream to DRAM on store.